chore: move annotation types to their own file
