+++ packages/playground/src/index.ts
@@ -2,3 +2,4 @@
 
 export { default as CogniteClientPlayground } from './cogniteClientPlayground';
 export * from './types';
+export * from './api/annotations/types';
